update runcoms to also work with folders
Now on the initialization the folder structure in the runcoms will also
be recreated in the users home folder (with a dot prefix). This will be
used to create a folder for a zshrc.d directory that will be traversed
for zshrc files. Then other repositories can create aliases and such.